move smap and remove_repeat to utils
