+++ src/mrpro/data/_KTrajectory.py
@@ -19,6 +19,8 @@ from dataclasses import dataclass
 import numpy as np
 import torch
 
+from mrpro.utils import remove_repeat
+
 
 @dataclass(slots=True, init=False)
 class KTrajectory:
@@ -70,7 +72,7 @@ class KTrajectory:
             Tolerance for repeat detection. Set to None to disable.
         """
         if repeat_detection_tolerance is not None:
-            kz, ky, kx = (KTrajectory._remove_repeat(tensor, repeat_detection_tolerance) for tensor in (kz, ky, kx))
+            kz, ky, kx = (remove_repeat(tensor, repeat_detection_tolerance) for tensor in (kz, ky, kx))
 
         self.kz = kz
         self.ky = ky
@@ -107,25 +109,3 @@ class KTrajectory:
 
         kz, ky, kx = torch.unbind(tensor, dim=stack_dim)
         return cls(kz, ky, kx, repeat_detection_tolerance=repeat_detection_tolerance)
-
-    @staticmethod
-    def _remove_repeat(tensor: torch.Tensor, tol: float) -> torch.Tensor:
-        """Replace dimensions with all equal values with singletons.
-
-        Parameters
-        ----------
-        tensor:
-            The tensor. Must be real
-        tol:
-            The tolerance
-        """
-        # TODO: Move to utilities
-
-        def can_be_singleton(dim: int) -> bool:
-            # If the distance between min and max is smaller than the tolerance, all values are the same.
-            return bool(torch.all((tensor.amax(dim=dim) - tensor.amin(dim=dim)) <= tol).item())
-
-        take_first = slice(0, 1)
-        take_all = slice(None)
-        index = tuple(take_first if can_be_singleton(dim) else take_all for dim in range(tensor.ndim))
-        return tensor[index]
